LayoutTests/imported/w3c/web-platform-tests/requestidlecallback/callback-exception-expected.txt

CONSOLE MESSAGE: Error: requestIdleCallbackException

PASS requestIdleCallback callback exceptions are reported to error handler
